+++ flake.nix
@@ -174,7 +174,7 @@
             installScriptFiles = nix-tool-installer.lib.${system}.mkInstallScriptFiles {
               toolName = "garn";
               baseUrl = "https://garn.io";
-              flakeLocation = "github:garnix-io/garn/v0.0.18";
+              flakeLocation = "github:garnix-io/garn/v0.0.19";
               testCommand = "garn --help";
             };
           }